Removes TODO in the wrong place.
The TODO in MMIMove::applicable_submatrices was put where the bug first
originated, however that is not where the bug lies at all. There is now a test
in test/mmi_move_test.cc that reproduces the issue, along with a TODO tag there.
That is a better place to have it, so this TODO tag was removed.