Adds SSMove class for sink-source moves.
Sink source mutations do not change the arrows in the quiver, except their
direction, so these preserve MMI-ness. However they cannot be checked using
standard MMI Moves, as they do no require a complete subquiver to be embedded.
Instead I have created a special move to handle this case. This will then be
added to the MoveClassLoader, so that when the class is computed, sink-source
mutations are also considered.