Adds class to construct graph of all moves.
The MoveGraph class contructs the whole class which is move-equivalent to the
initial matrix. This class is then stored in an unordered_map where each matrix
points to a 'Link' holding all matrices adjacent to it.

The QuiverGraph class is not ready yet. It needs to include a check in the
seen_matrix method to ensure that the links are only added once and needs to be
checked that everything is correct.

Both currently have no tests, so these need to be created.