Adds checks to only add quivers to graph if needed.
The seen_matrix method now only adds links if the new matrix exactly matches one
already seen. This ensures that the map links contain the expected matrices,
rather than permutations of them.

Also makes changes to Makefile. When building on Cygwin the -fPIC flag is not
rquired, so added some checks to remove the warnings generated by this. It turns
out that Cygwin's gcc stdlib has a bug where std::stoi is removed for some
reason, so the code will not compile anyway.

There seem to be some work arounds available by modifying the std headers, but
I could not get this to work.