Adds a disabled test to print out the moves in the class.
To confirm that the loader is working as expected the qvmove consts are added to
the library and these are used by a test which simply prints out all moves from
an initial matrix. This can be used to check that the results from qvmove make
sense.

It is kind of stupid having this as a test. Perhaps it would be better as a
utility in qvmove or something similar.