Changes MoveGraph to consider each node up to sink-source.
Sink source mutations are a type of move, but for the drawing of the move graph
they tend to confuse things, with many similar looking nodes in the graph. To
simplify the graph and better show the moves which are actually interesting this
commit changes the behaviour to consider the node up to sink-source mutations.
So the node can be thought of as a class of quivers which differ only by
sink-source muts.