Moves exchange graph specialisations into header.
Instead of having partial specialisations for every permutation of seed
and exchange graph we now have a number of template enable_if functions
which get enabled to provide the correct methods for any type of exchange
graph needed so far.

This means that any new graphs (eg sink source graphs or whatever) do not
need to specify anything other than an operator which specifies whether to
mutate a given matrix at a given vertex.