Reworking and optimizing the main check flow.
The main performance difference comes from not checking that the
row sums are permutations before computing the row mappings.
Previously, by doing this check early, we were essentially computing
the same thing twice, so as to avoid a small amount of other
computation. Now we pay this small price in some cases, but avoid the
large cost of computing the permutation problem twice.

Also changes some of the private methods and inlines others.

Uses pointers to rows rather than copying rows in do_rows_match
so the cached vectors are not needed. This reduced the amount of
copying between matrices and the checker.